testbuild

+++ tests/utils.py
@@ -222,7 +222,7 @@ schema = Schema(
         ),
         "total_profit_pct": Or(float, 0),
         "is_decimal": bool,
-        "rank": Or(int, None),
+        "rank": Or(float, None),
         "daily_live_profit": Or(float, 0),
         "total_user_invested_amount": Or(float, 0),
     }
